ai-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pydantic-settings
cachetools
sentence-transformers